requirements.txt
# streamlit==1.36.0
# langchain==0.2.5
# langchain-groq==0.1.5
# langchain-core==0.2.7
# langchain-huggingface==0.0.3
# langchain-pinecone==0.1.1
# langchain_openai==0.1.8
# pinecone-client==3.2.2
# langchain-community==0.2.5
# python-dotenv==1.0.1
# pypdf==4.2.0
# chromadb==0.5.3
# protobuf==4.25.3

streamlit==1.36.0
langchain
langchain-groq
langchain-core
langchain-huggingface
langchain-pinecone
pinecone-client
langchain-community
python-dotenv
pypdf
chromadb==0.5.3
langchain-chroma
pysqlite3-binary
pyodbc
# protobuf==3.20.3
# pydantic<2.0.0

